LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-handler-null-or-undefined-expected.txt


Harness Error (TIMEOUT), message = null

PASS event.intercept() should throw if the handler is null
TIMEOUT event.intercept() should not throw if the handler is explicit undefined Test timed out
